requirements.txt
Flask
python-dotenv
requests
flask-cors
